requirements.txt
django==3.2
django-cors-headers==3.12.0
djangorestframework==3.13.1
python-dotenv==0.19.2
django-extensions==3.1.5
django-filter==21.1
psycopg2-binary==2.9.3
Pillow==9.2.0
drf-writable-nested==0.7.0
